portfolio-backend/requirements.txt
asgiref==3.5.2
astroid==2.3.3
boto3==1.26.16
botocore==1.29.16
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.3
cryptography==39.0.1
Django==4.1.7
django-cors-headers==3.11.0
django-health-check==3.17.0
django-storages==1.13.1
djangorestframework==3.14.0
docutils==0.15.2
gunicorn==20.1.0
http-ece==1.1.0
idna==3.4
isort==4.3.21
jmespath==0.9.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==9.3.0
psycopg2==2.9.5
py-vapid==1.9.0
pycparser==2.21
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2022.1
pywebpush==1.14.0
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.13
wrapt==1.11.2
